rkt: Rename rocket to rkt.
